# docs/swagger.yaml
definitions:
  domain.Appointment:
    properties:
      date_and_time:
        type: string
      dentist_license:
        type: string
      description:
        type: string
      id:
        type: integer
      patient_identity:
        type: string
    required:
    - date_and_time
    - dentist_license
    - description
    - patient_identity
    type: object
...
  web.errorResponse:
    properties:
      message:
        type: string
      status:
        type: string
      status_code:
        type: integer
      time_stamp:
        type: string
    type: object
...
info:
  contact:
    name: Marcelo Ramos
    url: https://github.com/rows15
  description: This API handle appointments, dentists and patients for dental clinic
    system.
  license:
    name: Apache 2.0
    url: https://www.apache.org/license/LICENSE-2.0.html
  termsOfService: https://www.linkedin.com/in/rows15
  title: Dental Clinic API
  version: "1.0"
paths:
  /appointments:
    get:
      consumes:
      - application/json
      description: get all appointments from db.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Appointment'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
      security:
      - SECRET_TOKEN: []
      summary: List all appointments
      tags:
      - Appointments
    post:
      consumes:
      - application/json
      description: Create a new appointment by request body
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.Appointment'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Appointment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.errorResponse'
      security:
      - SECRET_TOKEN: []
      summary: Create a new appointment
      tags:
      - Appointments
  /appointments/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an appointment by ID
      parameters:
      - description: Appointment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.errorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
      security:
      - SECRET_TOKEN: []
      summary: Delete an appointment by ID
      tags:
      - Appointments
    get:
      consumes:
      - application/json
      description: get an appointment by a provided ID.
      parameters:
      - description: Appointment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Appointment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
      security:
      - SECRET_TOKEN: []
      summary: Get an appointment by an ID
      tags:
      - Appointments
    patch:
      consumes:
      - application/json
      description: Update fields from an appointment by ID and fields in request body
      parameters:
      - description: Appointment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.Appointment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Appointment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
      security:
      - SECRET_TOKEN: []
      summary: Update fields from an appointment by ID
      tags:
      - Appointments
    put:
      consumes:
      - application/json
      description: Update a entire appointment by ID and request body
      parameters:
      - description: Appointment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.Appointment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Appointment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.errorResponse'
      security:
      - SECRET_TOKEN: []
      summary: Update a entire appointment by ID
      tags:
      - Appointments
  /appointments/dentist/{license_number}:
    get:
      consumes:
      - application/json
      description: Get all appointments by dentist license doc
      parameters:
      - description: Dentist License Number
        in: path
        name: license_number
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Appointment'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
      security:
      - SECRET_TOKEN: []
      summary: Get all appointments by dentist license doc
      tags:
      - Appointments
  /appointments/patient/{identity_number}:
    get:
      consumes:
      - application/json
      description: get all appointments by patient identity doc
      parameters:
      - description: Patient Doc Number
        in: path
        name: identity_number
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Appointment'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
      security:
      - SECRET_TOKEN: []
      summary: Get all appointments by patient identity doc
      tags:
      - Appointments
...
securityDefinitions:
  SECRET_TOKEN:
    in: header
    name: SECRET_TOKEN
    type: apiKey
swagger: "2.0"